backend/pyproject.toml
[project]
name = "intelligent-job-management"
version = "0.1.0"
description = "Intelligent Job Management System for GPU-Accelerated Deep Learning Clusters"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.0",
    "uvicorn[standard]>=0.41.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.13.0",
    "psycopg[binary]>=3.3.0",
    "psycopg-pool>=3.2.0",
    "nats-py>=2.14.0",
    "python-multipart>=0.0.22",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
    "mypy>=1.19.0",
    "deptry>=0.12.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0.0",
    "types-pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/**/*.py" = ["ARG", "S101"]  # allow asserts and unused args in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", ".."]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
]

[tool.mypy]
python_version = "3.13"
mypy_path = ".."
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
